+++ CHANGELOG.rst
@@ -1,7 +1,51 @@
-7.0.0a1 (unreleased)
+7.0.0a6 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- ElasticSearchUtility: removed obsolete methods get_by_uuid(), get_by_uuids(),
+  get_by_type() and get_by_type()
+  [masipcat]
+- ElasticSearchUtility: removed unused internal method _get_type_query()
+  [masipcat]
+
+
+7.0.0a5 (2021-07-30)
+--------------------
+
+- Modifying the parse to be able to parse the __or field as a list of elements. Eg:
+  catalog_utility.search({"type_name__or": ["Folder", "Item"]})
+  [nilbacardit26]
+
+
+7.0.0a4 (2021-07-15)
+--------------------
+
+- Adding path as a wildcard in the parser. Searches will be done
+  within the same context using the endpoint @search. FYI: If no depth is
+  specified, the query resolves greater or equal than the content depth of the context plus one
+  [nilbacardit26]
+
+
+7.0.0a3 (2021-06-22)
+--------------------
+
+- Fixing the date parser. Using the default format
+  strict_date_optional_time instead of using the epoch_millis that
+  leads to the error: failed to parse date field [1.624173663E9].
+
+
+7.0.0a2 (2021-06-09)
+--------------------
+
+- Added __or in parser. With this we can query keywords with the or
+  operator like: `type_name=Item&__or=id=foo_item%26id=foo_item2`
+  [nilbacardit26]
+
+
+7.0.0a1 (2021-06-04)
+--------------------
+
+- Unrestricted parameter added to search_raw, in order to search as root
+  [nilbacardit26]
 
 
 7.0.0a0 (2021-05-03)